TEC-3592 - change naming, add in initialization of view coding, update doc blocks, update test
