+++ src/views/v2/widget-list.php
@@ -1,0 +1,8 @@
+<?php
+/*
+ * Temporary Display of V2 List Widget
+ */
+?>
+<div>
+	<?php echo "Widget Name: {$name}<br>"; ?>
+</div>
