+++ tests/views_widgets/Tribe/Events/Views/V2/Widgets/Widget_ListTest.php
@@ -6,7 +6,7 @@ use Tribe\Events\Views\V2\View;
 use Tribe\Events\Views\V2\Widgets\List_Widget;
 use Tribe\Test\Products\WPBrowser\Views\V2\ViewTestCase;
 
-class List_WidgetTest extends ViewTestCase {
+class Widget_ListTest extends ViewTestCase {
 
 	use MatchesSnapshots;
 
@@ -22,17 +22,7 @@ class List_WidgetTest extends ViewTestCase {
 		// Sanity check
 		$this->assertEmpty( tribe_events()->found() );
 
-		$context = tribe_context()->alter(
-			[
-				'event_display'      => 'list',
-				'event_display_mode' => 'list',
-				'today'              => $this->mock_date_value,
-				'now'                => $this->mock_date_value,
-				'event_date'         => $this->mock_date_value,
-			]
-		);
-
-		$list_view = View::make( List_Widget::class, $context );
+		$list_view = View::make( Widget_List_View::class );
 		$html      = $list_view->get_html();
 
 		// Let's make sure the View is displaying what events we expect it to display.
